requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.2+cpu
opencv-python-headless==4.5.5.62
absl-py==1.0.0
addict==2.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
filterpy==1.4.5
Flask==2.0.2
fonttools==4.29.1
future==0.18.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gunicorn==20.0.4
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.2
itsdangerous==2.1.0
Jinja2==3.0.3
kiwisolver==1.3.2
llvmlite==0.38.0
lmdb==1.3.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
networkx==2.7
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tb-nightly==2.9.0a20220303
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.13.1
tifffile==2022.2.9
tornado==6.1
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
yapf==0.32.0
zipp==3.7.0